feat: Add register and verify otp email register page
refactor: delete back-button, update btn which pages are using, add .env in constantts file
